+++ db/revert/util_functions/migrate_dependencies.sql
@@ -1,0 +1,7 @@
+-- Revert ccbc:util_functions/migrate_dependencies from pg
+
+BEGIN;
+
+-- XXX Add DDLs here.
+
+COMMIT;